bibliotecario-parcial/bibliotecario/requirements.txt
chatterbot
chatterbot_corpus
nltk # natural language toolkit

flask
pytesseract #instalar também o pacote OCR para o idioma portugues: sudo apt install tesseract-ocr-por
pillow
transformers
torch

langchain_google_genai
protobuf
